pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "xl2times-mcp-server"
version = "0.1.0"
description = "MCP server for xl2times VEDA-TIMES processing"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp>=0.2.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "fastapi>=0.115.0",
]

[project.scripts]
xl2times-mcp-server = "src.server:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]  # Line too long - handled by formatter

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"

[tool.hatch.build.targets.wheel]
packages = ["src"]
